Make the pytest fixtures execute once but print everywhere
The problem here was that run_task() printed.
By default, pytest would only print the captured output on a failed test.

If we make the fixture session scoped and the first test using this
fixture passed, and another one failed, it would have not be printed.

This is a function scoped fixture that uses a session scoped fixture
to workaround this problem.

Now we save time because we only execute the task once, but the prints
are preserved.