Make sure Bugzilla is queried without being logged in
After a few months of not running the integration tests
I ended up with a strange error message:

<Fault 32000: 'The cookies or token provide were not valid or have
expired. You may login again to get new cookies or a new token.'>

Apparently, bugzilla.Bugzilla(...) is logging me in with some
credentials from god-knows-where. It worked last time, but in the
meantime, those credentials have expired.

I found a discussion about this at [1], and went with a solution
that disables this automatic logging in, keeping the Bugzilla
query anonymous. That fixed my issue.

[1]: https://lists.fedorahosted.org/archives/list/python-bugzilla@lists.fedorahosted.org/message/A2ZUGBFWATW676YRPJMTL2F2WYWT5RMB/